Improved data requirement docs (#67)
Note added to data requirements specifically covering mismatching of headers when data is split between files for a single omop target.
Additional notes on the person file added, specifically about its role as a registry, what happens if person ids are not in it, and that it can also be used as a data input.